requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.2
pydantic>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
ultralytics>=8.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
lap>=0.5.12
